requirements.txt
Flask==3.1.0
Flask-Cors==5.0.1
PyJWT==2.10.1
bcrypt==4.3.0
mysql-connector-python==9.2.0
gunicorn==22.0.0
flask-jwt-extended==4.7.1
python-dotenv==1.0.1
cryptography==41.0.0